Script should be debugged with common functions.
